Cleanup by moving each class into a separate file.  Public / Open classes retain MJ prefix for now even though that is a bit of an objc thing.
