+++ MJFastImageLoader/MJFastImageLoader/MJFastImageLoader.swift
@@ -230,102 +230,6 @@ public class MJFastImageLoader {
 		leastRecentlyUsed = []
 	}
 
-	open class MJFastImageLoaderNotification: Equatable {
-		// Use a linked list because the most likely cases will be zero or one node, and multi-node won't involve searching.
-		var next:MJFastImageLoaderNotification? = nil
-		var cancelled = false
-		var workItem:WorkItem? = nil
-		private let batch:MJFastImageLoaderBatch?
-
-		public init(batch: MJFastImageLoaderBatch?) {
-			self.batch = batch
-		}
-
-		func queueNotifyEvent(image: UIImage) {
-			if let batch = batch {
-				batch.queueNotifyEvent(image: image, notification: self)
-			}
-			else {
-				notify(image: image)
-			}
-		}
-
-		open func notify(image: UIImage) {
-		}
-
-		open func cancel() {
-			cancelled = true
-			_ = workItem?.release() // for our retain
-		}
-
-		public static func == (lhs: MJFastImageLoaderNotification, rhs: MJFastImageLoaderNotification) -> Bool {
-			return lhs === rhs
-		}
-	}
-
-	open class MJFastImageLoaderBatch {
-		public static let shared = MJFastImageLoaderBatch()
-
-		open var batchUpdateQuantityLimit = 1
-		public var batchUpdateTimeLimit = 0.1
-		
-		private var notifications:[MJFastImageLoaderNotification] = []
-		private var images:[UIImage] = []
-		private let queue = DispatchQueue(label: "MJFastImageLoaderBatch.queue")
-		private var timeLimitWorkItem:DispatchWorkItem? = nil
-
-		func queueNotifyEvent(image: UIImage, notification: MJFastImageLoaderNotification) {
-			queue.sync {
-				if let index = notifications.index(of: notification) {
-					images[index] = image
-				}
-				else {
-					notifications.append(notification)
-					images.append(image)
-				}
-
-				var nonCancelledCount = 0
-				for i in 0..<notifications.count {
-					if ( !notifications[i].cancelled ) {
-						nonCancelledCount += 1
-					}
-				}
-
-				let first = 1 == nonCancelledCount
-				let hitQuota = nonCancelledCount >= batchUpdateQuantityLimit
-				if ( first ) {
-					// In case the previous timer were still running for content that had all been cancelled.
-					timeLimitWorkItem?.cancel()
-					timeLimitWorkItem = nil
-				}
-				if ( first || hitQuota ) {
-					let timeout:Double = hitQuota ? 0.0 : batchUpdateTimeLimit
-					//let timeout2:Int = hitQuota ? 0 : 5
-					timeLimitWorkItem = DispatchWorkItem {
-						// Cancel the timeout, if there were one
-						self.timeLimitWorkItem?.cancel()
-						self.timeLimitWorkItem = nil
-
-						NSLog("notify \(self.notifications.count) for \(hitQuota ? "quota" : "timeout")")
-
-						// Do our work
-						DispatchQueue.main.sync {
-							// Use the main queue so that batches will draw as one.  Yes, this matters.
-							for i in 0..<self.notifications.count {
-								if ( !self.notifications[i].cancelled ) {
-									self.notifications[i].notify(image: self.images[i])
-								}
-							}
-						}
-						self.notifications = []
-						self.images = []
-					}
-					queue.asyncAfter(deadline: .now() + timeout, execute: timeLimitWorkItem!)
-				}
-			}
-		}
-	}
-
 	// MARK: Private Variables - Settings
 
 	var criticalProcessingConcurrencyLimit = 12
@@ -339,167 +243,6 @@ public class MJFastImageLoader {
 	var maxResultsVolumeBytes = 0
 	var leastRecentlyUsed:[Data] = []
 	var hintMap:[Data:Int] = [:]
-
-	class WorkItem {
-		init(data: Data, uid: Int, basePriority: Priority) {
-			self.data = data
-			self.uid = uid
-			self.basePriority = basePriority
-		}
-
-		var priority: Int {
-			// Only add state to decrease priority if we have rendered something already
-			return basePriority.rawValue + (haveImage ? state : 0)
-		}
-
-		let data:Data
-		let uid:Int
-		var basePriority:Priority
-		var isCancelled = false
-		var state:Int = 0
-		var currentImage:UIImage? = nil
-		var haveImage = false
-		var notification:MJFastImageLoaderNotification? = nil
-
-		public static let retainQueue = DispatchQueue(label: "MJFastImageLoader.workItemRetention")
-		var retainCount = 1
-
-		public func retain () {
-			WorkItem.retainQueue.sync {
-				self.retainCount += 1
-			}
-		}
-		public func release () -> Bool {
-			var nonZero = true
-			WorkItem.retainQueue.sync {
-				self.retainCount -= 1
-				nonZero = self.retainCount > 0
-			}
-			return nonZero
-		}
-
-		public func next( thumbnailPixels: Float ) -> UIImage? {
-			let thumbnailMaxPixels = thumbnailPixels
-			let cgThumbnailMaxPixels = CGFloat(thumbnailPixels)
-
-			print("state \(state)")
-			switch state {
-			case 0:
-				// Fastest first.  Use thumbnail if present.
-				state += 1
-
-				let imageSource = CGImageSourceCreateWithData(data as CFData, nil)
-
-				let options: CFDictionary = [
-					kCGImageSourceShouldAllowFloat as String: true as NSNumber,
-					kCGImageSourceCreateThumbnailWithTransform as String: true as NSNumber,
-					kCGImageSourceCreateThumbnailFromImageIfAbsent as String: true as NSNumber,
-					kCGImageSourceThumbnailMaxPixelSize as String: thumbnailMaxPixels as NSNumber
-					] as CFDictionary
-
-				if let thumbnail = CGImageSourceCreateThumbnailAtIndex(imageSource!, 0, options)
-				{
-					let result = UIImage(cgImage: thumbnail)
-
-					if ( result.size.width >= cgThumbnailMaxPixels
-						|| result.size.height >= cgThumbnailMaxPixels )
-					{
-						state += 1
-					}
-
-					currentImage = result
-					haveImage = true
-					notify(notification: notification, image: result, previous: nil)
-					return result
-				}
-
-				return next(thumbnailPixels: thumbnailPixels) // Immediately provide next image if we couldn't provide this one.
-
-			case 1:
-				// Generate better thumbnail if appropriate.
-				state += 1
-
-				let imageSource = CGImageSourceCreateWithData(data as CFData, nil)
-
-				let options: CFDictionary = [
-					kCGImageSourceShouldAllowFloat as String: true as NSNumber,
-					kCGImageSourceCreateThumbnailWithTransform as String: true as NSNumber,
-					kCGImageSourceCreateThumbnailFromImageAlways as String: true as NSNumber,
-					kCGImageSourceThumbnailMaxPixelSize as String: thumbnailMaxPixels as NSNumber
-					] as CFDictionary
-
-				if let thumbnail = CGImageSourceCreateThumbnailAtIndex(imageSource!, 0, options)
-				{
-					let result = UIImage(cgImage: thumbnail)
-					currentImage = result
-					haveImage = true
-					notify(notification: notification, image: result, previous: nil)
-					return result
-				}
-
-				return next(thumbnailPixels: thumbnailPixels) // Immediately provide next image if we couldn't provide this one.
-
-			case 2:
-				// Generate final image last.
-				state += 1
-
-				let previousImage = currentImage
-				currentImage = nil // Since we won't need this after this stage.
-
-				if let image = UIImage(data: data) {
-					if ( image.size == previousImage?.size )
-					{
-						// Prior call produced full-size image, so stop now.
-						return nil
-					}
-
-					UIGraphicsBeginImageContextWithOptions(image.size, false, image.scale)
-					image.draw(at: .zero)
-					let resultImage = UIGraphicsGetImageFromCurrentImageContext()
-					UIGraphicsEndImageContext()
-					if let resultImage = resultImage
-					{
-						notify(notification: notification, image: resultImage, previous: nil)
-					}
-					if ( nil != resultImage ) {
-						haveImage = true
-					}
-					return resultImage
-				}
-
-				return nil
-
-			default:
-				return nil
-			}
-		}
-
-		func notify(notification: MJFastImageLoaderNotification?, image: UIImage, previous: MJFastImageLoaderNotification?) {
-			// Handle the linked list ourselves so it is not vulnerable to breakage by implementors of items in it
-			if ( nil == notification && nil == previous )
-			{
-				print("old nobody cares")
-				MJFastImageLoader.wasteCount += 1
-			}
-			if let notification = notification {
-				if ( notification.cancelled ) {
-					// Clear out cancelled item
-					if let previous = previous {
-						previous.next = notification.next
-					}
-					else {
-						self.notification = notification.next
-					}
-				}
-				else {
-					notification.queueNotifyEvent(image: image)
-				}
-
-				// Notify next link in the list
-				notify(notification: notification.next, image: image, previous: notification)
-			}
-		}
-	}
 
 	func processWorkItem() {
 // fixme - wrap this all in a non-concurrent GCD queue
